[GPU] Refactoring
