+++ src/plugins/intel_gpu/include/intel_gpu/graph/kernel_impl_params.hpp
@@ -24,6 +24,12 @@ struct program;
 
 
 struct kernel_impl_params {
+    struct Hasher {
+        size_t operator()(const kernel_impl_params &k) const {
+            return k.hash();
+        }
+    };
+
     bool has_runtime_layouts = false;
     const program *prog;
     std::shared_ptr<const primitive> desc;
@@ -116,53 +122,8 @@ struct kernel_impl_params {
         return *prog;
     }
 
-    size_t hash() const {
-        size_t seed = desc->hash();
-        const size_t prime_number = 2654435761; // magic number to reduce hash collision rate.
-        for (auto& in : input_layouts) {
-            seed = hash_combine(seed, in.hash() * prime_number);
-        }
-        for (auto& out : output_layouts) {
-            seed = hash_combine(seed, out.hash() * prime_number);
-        }
-
-        // hash for fused prims
-        for (auto& fd : fused_desc) {
-            seed = hash_combine(seed, fd.desc->hash());
-        }
-        return seed;
-    }
-
-    bool operator==(const kernel_impl_params& rhs) const {
-        if (*desc != *rhs.desc)
-            return false;
-
-        if (rhs.input_layouts.size() != input_layouts.size())
-            return false;
-
-        if (rhs.output_layouts.size() != output_layouts.size())
-            return false;
-
-        for (size_t i = 0; i < input_layouts.size(); i++) {
-            if (input_layouts[i] != rhs.input_layouts[i])
-                return false;
-        }
-
-        for (size_t i = 0; i < output_layouts.size(); i++) {
-            if (output_layouts[i] != rhs.output_layouts[i])
-                return false;
-        }
-
-        if (fused_desc.size() != rhs.fused_desc.size())
-            return false;
-
-        for (size_t i = 0; i < rhs.fused_desc.size(); i++) {
-            if (fused_desc[i] != rhs.fused_desc[i])
-                return false;
-        }
-
-        return true;
-    }
+    size_t hash() const;
+    bool operator==(const kernel_impl_params& rhs) const;
 };
 
 }  // namespace cldnn
