+++ src/plugins/intel_gpu/include/intel_gpu/graph/program.hpp
@@ -225,7 +225,6 @@ public:
 
     void add_optimized_primitive_info(primitive_id optimized_primitive_id, std::vector<primitive_id> replaced_with_ids = {});
 
-    void reset_program();
     uint32_t get_id() const { return prog_id; }
 
     static ptr build_program(engine& engine,
@@ -240,8 +239,6 @@ public:
                              std::shared_ptr<InferenceEngine::CPUStreamsExecutor> task_executor,
                              bool is_internal);
     static void init_primitives();
-    void compile();
-    void init_kernels();
     kernel_id add_kernel(const std::shared_ptr<kernel_string>& kernel_sring);
     kernel::ptr get_kernel(kernel_id id);
     kernels_cache& get_kernels_cache() const;
@@ -251,13 +248,7 @@ public:
 
     void remove_kernel(kernel_id id);
 
-    struct ImplHasher {
-        size_t operator()(const kernel_impl_params &k) const {
-            return k.hash();
-        }
-    };
-
-    using ImplementationsCache = cldnn::LruCacheThreadSafe<kernel_impl_params, std::shared_ptr<primitive_impl>, ImplHasher>;
+    using ImplementationsCache = cldnn::LruCacheThreadSafe<kernel_impl_params, std::shared_ptr<primitive_impl>, kernel_impl_params::Hasher>;
     ImplementationsCache& get_implementations_cache() const { return *_impls_cache; }
     ICompilationContext& get_compilation_context() const { return *_compilation_context; }
     void cancel_compilation_context();
@@ -314,7 +305,6 @@ private:
     void run_graph_compilation();
     void pre_optimize_graph(bool is_internal);
     void post_optimize_graph(bool is_internal);
-    void cleanup();
     void transfer_memory_to_device();
 
     InferenceEngine::CPUStreamsExecutor::Config make_task_executor_config(const ExecutionConfig& config, std::string tags = "") const;
