+++ src/plugins/intel_gpu/src/graph/include/eltwise_inst.h
@@ -11,57 +11,6 @@
 
 namespace cldnn {
 
-inline kernel_selector::eltwise_mode convert_to_eltwise_mode(eltwise_mode mode) {
-    switch (mode) {
-        case eltwise_mode::sum:
-            return kernel_selector::eltwise_mode::ADD;
-        case eltwise_mode::sub:
-            return kernel_selector::eltwise_mode::SUB;
-        case eltwise_mode::max:
-            return kernel_selector::eltwise_mode::MAX;
-        case eltwise_mode::prod:
-            return kernel_selector::eltwise_mode::MUL;
-        case eltwise_mode::div:
-            return kernel_selector::eltwise_mode::DIV;
-        case eltwise_mode::min:
-            return kernel_selector::eltwise_mode::MIN;
-        case eltwise_mode::pow:
-            return kernel_selector::eltwise_mode::POW;
-        case eltwise_mode::mod:
-            return kernel_selector::eltwise_mode::MODULU;
-        case eltwise_mode::eq:
-            return kernel_selector::eltwise_mode::EQ;
-        case eltwise_mode::ne:
-            return kernel_selector::eltwise_mode::NE;
-        case eltwise_mode::lt:
-            return kernel_selector::eltwise_mode::LT;
-        case eltwise_mode::le:
-            return kernel_selector::eltwise_mode::LE;
-        case eltwise_mode::gt:
-            return kernel_selector::eltwise_mode::GT;
-        case eltwise_mode::ge:
-            return kernel_selector::eltwise_mode::GE;
-        case eltwise_mode::logic_and:
-            return kernel_selector::eltwise_mode::LOGIC_AND;
-        case eltwise_mode::logic_or:
-            return kernel_selector::eltwise_mode::LOGIC_OR;
-        case eltwise_mode::logic_xor:
-            return kernel_selector::eltwise_mode::LOGIC_XOR;
-        case eltwise_mode::squared_diff:
-            return kernel_selector::eltwise_mode::SQUARED_DIFF;
-        case eltwise_mode::floor_mod:
-            return kernel_selector::eltwise_mode::FLOOR_MOD;
-        case eltwise_mode::is_finite:
-            return kernel_selector::eltwise_mode::IS_FINITE;
-        case eltwise_mode::is_inf:
-            return kernel_selector::eltwise_mode::IS_INF;
-        case eltwise_mode::is_nan:
-            return kernel_selector::eltwise_mode::IS_NAN;
-        default:
-            return kernel_selector::eltwise_mode::ADD;
-    }
-}
-
 class EltwiseFuseParams : public NodeFuseParams {
 public:
     EltwiseFuseParams(std::shared_ptr<eltwise> desc) : NodeFuseParams(eltwise::type_id()), _desc(desc) {}
