+++ src/plugins/intel_gpu/tests/passes/prepare_buffer_fusing_test.cpp
@@ -86,8 +86,7 @@ TEST(prepare_buffer_fusing, static_node_after_optimized_out_dyn_reshape) {
     program_wrapper::apply_opt_pass<compile_graph>(*prog);
     ASSERT_NO_THROW(prog->get_node("reshape"));
     ASSERT_FALSE(prog->get_node("reshape").can_be_optimized());
-    prog->compile();
-    prog->init_kernels();
+    program_wrapper::apply_opt_pass<build_implementations>(*prog);
 
     ASSERT_TRUE(has_node_with_type<reshape>(*prog));
